abc138 q2 thinking

+++ AtCoderProblems/abc138/Alchemist.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int main(){
     int N;
     cin >> N;
-    vector<int> v(N);
+    vector<double> v(N);
 
     for(int i = 0; i < N; i++){
         cin >> v.at(i);
